requirements.txt
signalbot>=0.16.0
python-dateutil>=2.8.0
pyyaml>=6.0
requests>=2.28.0

